Age pyramid from a population data frame at a given time.

Reduce a population data frame containing all individuals (with some characteristics) to an age-groups data frame (preserving characteristics). The function computes the number of individuals at time in each age group [ages[i],ages[i+1][, for i in {1,...,N-1}.

age_pyramid(population, time = 0, ages = c(0:110, Inf))

Arguments

population

Population data frame: characteristics in columns and individuals in rows. At least two columns birth (dates of birth) and death (dates of death) are required.

time

The age pyramid is computed at instant time. Must be a numeric greater than or equal to 0.

ages

(Optional) A numeric vector of distinct positive values composing age groups. Must be in increasing order.

Value

Age pyramid of a population at instant time. A data frame with columns age and value. Optional characteristics are preserved.

Details

See also age_pyramids.

Examples

age_pyramid(EW_pop_14$sample, time = 0)
#>           age  male value
#> 1       0 - 1 FALSE   581
#> 2       1 - 2 FALSE   577
#> 3       2 - 3 FALSE   638
#> 4       3 - 4 FALSE   632
#> 5       4 - 5 FALSE   590
#> 6       5 - 6 FALSE   570
#> 7       6 - 7 FALSE   579
#> 8       7 - 8 FALSE   588
#> 9       8 - 9 FALSE   548
#> 10     9 - 10 FALSE   532
#> 11    10 - 11 FALSE   503
#> 12    11 - 12 FALSE   562
#> 13    12 - 13 FALSE   514
#> 14    13 - 14 FALSE   509
#> 15    14 - 15 FALSE   573
#> 16    15 - 16 FALSE   582
#> 17    16 - 17 FALSE   536
#> 18    17 - 18 FALSE   565
#> 19    18 - 19 FALSE   618
#> 20    19 - 20 FALSE   616
#> 21    20 - 21 FALSE   623
#> 22    21 - 22 FALSE   650
#> 23    22 - 23 FALSE   714
#> 24    23 - 24 FALSE   671
#> 25    24 - 25 FALSE   621
#> 26    25 - 26 FALSE   682
#> 27    26 - 27 FALSE   714
#> 28    27 - 28 FALSE   687
#> 29    28 - 29 FALSE   712
#> 30    29 - 30 FALSE   723
#> 31    30 - 31 FALSE   696
#> 32    31 - 32 FALSE   670
#> 33    32 - 33 FALSE   725
#> 34    33 - 34 FALSE   695
#> 35    34 - 35 FALSE   692
#> 36    35 - 36 FALSE   657
#> 37    36 - 37 FALSE   609
#> 38    37 - 38 FALSE   608
#> 39    38 - 39 FALSE   644
#> 40    39 - 40 FALSE   600
#> 41    40 - 41 FALSE   612
#> 42    41 - 42 FALSE   642
#> 43    42 - 43 FALSE   757
#> 44    43 - 44 FALSE   728
#> 45    44 - 45 FALSE   685
#> 46    45 - 46 FALSE   754
#> 47    46 - 47 FALSE   729
#> 48    47 - 48 FALSE   728
#> 49    48 - 49 FALSE   752
#> 50    49 - 50 FALSE   773
#> 51    50 - 51 FALSE   753
#> 52    51 - 52 FALSE   660
#> 53    52 - 53 FALSE   714
#> 54    53 - 54 FALSE   668
#> 55    54 - 55 FALSE   616
#> 56    55 - 56 FALSE   626
#> 57    56 - 57 FALSE   680
#> 58    57 - 58 FALSE   609
#> 59    58 - 59 FALSE   558
#> 60    59 - 60 FALSE   571
#> 61    60 - 61 FALSE   547
#> 62    61 - 62 FALSE   571
#> 63    62 - 63 FALSE   534
#> 64    63 - 64 FALSE   521
#> 65    64 - 65 FALSE   533
#> 66    65 - 66 FALSE   580
#> 67    66 - 67 FALSE   640
#> 68    67 - 68 FALSE   643
#> 69    68 - 69 FALSE   504
#> 70    69 - 70 FALSE   467
#> 71    70 - 71 FALSE   501
#> 72    71 - 72 FALSE   417
#> 73    72 - 73 FALSE   434
#> 74    73 - 74 FALSE   364
#> 75    74 - 75 FALSE   404
#> 76    75 - 76 FALSE   370
#> 77    76 - 77 FALSE   338
#> 78    77 - 78 FALSE   380
#> 79    78 - 79 FALSE   343
#> 80    79 - 80 FALSE   341
#> 81    80 - 81 FALSE   295
#> 82    81 - 82 FALSE   260
#> 83    82 - 83 FALSE   275
#> 84    83 - 84 FALSE   259
#> 85    84 - 85 FALSE   231
#> 86    85 - 86 FALSE   239
#> 87    86 - 87 FALSE   217
#> 88    87 - 88 FALSE   207
#> 89    88 - 89 FALSE   168
#> 90    89 - 90 FALSE   133
#> 91    90 - 91 FALSE   125
#> 92    91 - 92 FALSE   112
#> 93    92 - 93 FALSE    92
#> 94    93 - 94 FALSE    79
#> 95    94 - 95 FALSE    49
#> 96    95 - 96 FALSE    36
#> 97    96 - 97 FALSE    25
#> 98    97 - 98 FALSE    18
#> 99    98 - 99 FALSE    10
#> 100  99 - 100 FALSE    12
#> 101 100 - 101 FALSE     8
#> 102 101 - 102 FALSE     6
#> 103 102 - 103 FALSE     3
#> 104 103 - 104 FALSE     2
#> 105 104 - 105 FALSE     1
#> 106 105 - 106 FALSE     0
#> 107 106 - 107 FALSE     2
#> 108 107 - 108 FALSE     0
#> 109 108 - 109 FALSE     0
#> 110 109 - 110 FALSE     0
#> 111 110 - Inf FALSE     0
#> 112     0 - 1  TRUE   584
#> 113     1 - 2  TRUE   605
#> 114     2 - 3  TRUE   718
#> 115     3 - 4  TRUE   675
#> 116     4 - 5  TRUE   608
#> 117     5 - 6  TRUE   649
#> 118     6 - 7  TRUE   682
#> 119     7 - 8  TRUE   687
#> 120     8 - 9  TRUE   582
#> 121    9 - 10  TRUE   570
#> 122   10 - 11  TRUE   584
#> 123   11 - 12  TRUE   600
#> 124   12 - 13  TRUE   518
#> 125   13 - 14  TRUE   589
#> 126   14 - 15  TRUE   606
#> 127   15 - 16  TRUE   609
#> 128   16 - 17  TRUE   607
#> 129   17 - 18  TRUE   612
#> 130   18 - 19  TRUE   624
#> 131   19 - 20  TRUE   657
#> 132   20 - 21  TRUE   656
#> 133   21 - 22  TRUE   614
#> 134   22 - 23  TRUE   684
#> 135   23 - 24  TRUE   673
#> 136   24 - 25  TRUE   716
#> 137   25 - 26  TRUE   669
#> 138   26 - 27  TRUE   723
#> 139   27 - 28  TRUE   625
#> 140   28 - 29  TRUE   665
#> 141   29 - 30  TRUE   729
#> 142   30 - 31  TRUE   671
#> 143   31 - 32  TRUE   663
#> 144   32 - 33  TRUE   650
#> 145   33 - 34  TRUE   698
#> 146   34 - 35  TRUE   647
#> 147   35 - 36  TRUE   665
#> 148   36 - 37  TRUE   590
#> 149   37 - 38  TRUE   607
#> 150   38 - 39  TRUE   618
#> 151   39 - 40  TRUE   612
#> 152   40 - 41  TRUE   632
#> 153   41 - 42  TRUE   655
#> 154   42 - 43  TRUE   681
#> 155   43 - 44  TRUE   743
#> 156   44 - 45  TRUE   701
#> 157   45 - 46  TRUE   689
#> 158   46 - 47  TRUE   718
#> 159   47 - 48  TRUE   729
#> 160   48 - 49  TRUE   666
#> 161   49 - 50  TRUE   724
#> 162   50 - 51  TRUE   708
#> 163   51 - 52  TRUE   686
#> 164   52 - 53  TRUE   713
#> 165   53 - 54  TRUE   610
#> 166   54 - 55  TRUE   648
#> 167   55 - 56  TRUE   560
#> 168   56 - 57  TRUE   622
#> 169   57 - 58  TRUE   612
#> 170   58 - 59  TRUE   552
#> 171   59 - 60  TRUE   550
#> 172   60 - 61  TRUE   575
#> 173   61 - 62  TRUE   558
#> 174   62 - 63  TRUE   540
#> 175   63 - 64  TRUE   502
#> 176   64 - 65  TRUE   487
#> 177   65 - 66  TRUE   543
#> 178   66 - 67  TRUE   588
#> 179   67 - 68  TRUE   636
#> 180   68 - 69  TRUE   468
#> 181   69 - 70  TRUE   477
#> 182   70 - 71  TRUE   390
#> 183   71 - 72  TRUE   455
#> 184   72 - 73  TRUE   371
#> 185   73 - 74  TRUE   347
#> 186   74 - 75  TRUE   338
#> 187   75 - 76  TRUE   312
#> 188   76 - 77  TRUE   327
#> 189   77 - 78  TRUE   300
#> 190   78 - 79  TRUE   282
#> 191   79 - 80  TRUE   261
#> 192   80 - 81  TRUE   218
#> 193   81 - 82  TRUE   218
#> 194   82 - 83  TRUE   228
#> 195   83 - 84  TRUE   177
#> 196   84 - 85  TRUE   160
#> 197   85 - 86  TRUE   132
#> 198   86 - 87  TRUE   137
#> 199   87 - 88  TRUE   114
#> 200   88 - 89  TRUE   102
#> 201   89 - 90  TRUE    76
#> 202   90 - 91  TRUE    49
#> 203   91 - 92  TRUE    60
#> 204   92 - 93  TRUE    38
#> 205   93 - 94  TRUE    23
#> 206   94 - 95  TRUE    18
#> 207   95 - 96  TRUE    11
#> 208   96 - 97  TRUE     8
#> 209   97 - 98  TRUE     4
#> 210   98 - 99  TRUE     3
#> 211  99 - 100  TRUE     4
#> 212 100 - 101  TRUE     1
#> 213 101 - 102  TRUE     4
#> 214 102 - 103  TRUE     0
#> 215 103 - 104  TRUE     0
#> 216 104 - 105  TRUE     1
#> 217 105 - 106  TRUE     0
#> 218 106 - 107  TRUE     0
#> 219 107 - 108  TRUE     0
#> 220 108 - 109  TRUE     0
#> 221 109 - 110  TRUE     0
#> 222 110 - Inf  TRUE     0

age_pyramid(EW_popIMD_14$sample, time = 0, ages = seq(0, 120, by=2))
#>           age  male IMD value
#> 1       0 - 2 FALSE   1   166
#> 2       2 - 4 FALSE   1   203
#> 3       4 - 6 FALSE   1   219
#> 4       6 - 8 FALSE   1   255
#> 5      8 - 10 FALSE   1   203
#> 6     10 - 12 FALSE   1   192
#> 7     12 - 14 FALSE   1   235
#> 8     14 - 16 FALSE   1   239
#> 9     16 - 18 FALSE   1   250
#> 10    18 - 20 FALSE   1   213
#> 11    20 - 22 FALSE   1   186
#> 12    22 - 24 FALSE   1   193
#> 13    24 - 26 FALSE   1   200
#> 14    26 - 28 FALSE   1   171
#> 15    28 - 30 FALSE   1   157
#> 16    30 - 32 FALSE   1   198
#> 17    32 - 34 FALSE   1   227
#> 18    34 - 36 FALSE   1   255
#> 19    36 - 38 FALSE   1   216
#> 20    38 - 40 FALSE   1   238
#> 21    40 - 42 FALSE   1   272
#> 22    42 - 44 FALSE   1   304
#> 23    44 - 46 FALSE   1   315
#> 24    46 - 48 FALSE   1   304
#> 25    48 - 50 FALSE   1   307
#> 26    50 - 52 FALSE   1   338
#> 27    52 - 54 FALSE   1   289
#> 28    54 - 56 FALSE   1   274
#> 29    56 - 58 FALSE   1   246
#> 30    58 - 60 FALSE   1   255
#> 31    60 - 62 FALSE   1   233
#> 32    62 - 64 FALSE   1   240
#> 33    64 - 66 FALSE   1   260
#> 34    66 - 68 FALSE   1   333
#> 35    68 - 70 FALSE   1   245
#> 36    70 - 72 FALSE   1   239
#> 37    72 - 74 FALSE   1   164
#> 38    74 - 76 FALSE   1   158
#> 39    76 - 78 FALSE   1   152
#> 40    78 - 80 FALSE   1   143
#> 41    80 - 82 FALSE   1   142
#> 42    82 - 84 FALSE   1   121
#> 43    84 - 86 FALSE   1   107
#> 44    86 - 88 FALSE   1    83
#> 45    88 - 90 FALSE   1    58
#> 46    90 - 92 FALSE   1    49
#> 47    92 - 94 FALSE   1    48
#> 48    94 - 96 FALSE   1    21
#> 49    96 - 98 FALSE   1     9
#> 50   98 - 100 FALSE   1     4
#> 51  100 - 102 FALSE   1     2
#> 52  102 - 104 FALSE   1     6
#> 53  104 - 106 FALSE   1     0
#> 54  106 - 108 FALSE   1     0
#> 55  108 - 110 FALSE   1     0
#> 56  110 - 112 FALSE   1     0
#> 57  112 - 114 FALSE   1     0
#> 58  114 - 116 FALSE   1     0
#> 59  116 - 118 FALSE   1     0
#> 60  118 - 120 FALSE   1     0
#> 61      0 - 2  TRUE   1   203
#> 62      2 - 4  TRUE   1   243
#> 63      4 - 6  TRUE   1   219
#> 64      6 - 8  TRUE   1   256
#> 65     8 - 10  TRUE   1   208
#> 66    10 - 12  TRUE   1   212
#> 67    12 - 14  TRUE   1   205
#> 68    14 - 16  TRUE   1   237
#> 69    16 - 18  TRUE   1   259
#> 70    18 - 20  TRUE   1   213
#> 71    20 - 22  TRUE   1   224
#> 72    22 - 24  TRUE   1   205
#> 73    24 - 26  TRUE   1   202
#> 74    26 - 28  TRUE   1   183
#> 75    28 - 30  TRUE   1   171
#> 76    30 - 32  TRUE   1   177
#> 77    32 - 34  TRUE   1   221
#> 78    34 - 36  TRUE   1   183
#> 79    36 - 38  TRUE   1   212
#> 80    38 - 40  TRUE   1   224
#> 81    40 - 42  TRUE   1   255
#> 82    42 - 44  TRUE   1   255
#> 83    44 - 46  TRUE   1   278
#> 84    46 - 48  TRUE   1   315
#> 85    48 - 50  TRUE   1   293
#> 86    50 - 52  TRUE   1   302
#> 87    52 - 54  TRUE   1   307
#> 88    54 - 56  TRUE   1   265
#> 89    56 - 58  TRUE   1   280
#> 90    58 - 60  TRUE   1   257
#> 91    60 - 62  TRUE   1   218
#> 92    62 - 64  TRUE   1   236
#> 93    64 - 66  TRUE   1   259
#> 94    66 - 68  TRUE   1   269
#> 95    68 - 70  TRUE   1   219
#> 96    70 - 72  TRUE   1   192
#> 97    72 - 74  TRUE   1   159
#> 98    74 - 76  TRUE   1   158
#> 99    76 - 78  TRUE   1   126
#> 100   78 - 80  TRUE   1   107
#> 101   80 - 82  TRUE   1   120
#> 102   82 - 84  TRUE   1   107
#> 103   84 - 86  TRUE   1    83
#> 104   86 - 88  TRUE   1    56
#> 105   88 - 90  TRUE   1    44
#> 106   90 - 92  TRUE   1    24
#> 107   92 - 94  TRUE   1    14
#> 108   94 - 96  TRUE   1    12
#> 109   96 - 98  TRUE   1     2
#> 110  98 - 100  TRUE   1     1
#> 111 100 - 102  TRUE   1     0
#> 112 102 - 104  TRUE   1     1
#> 113 104 - 106  TRUE   1     0
#> 114 106 - 108  TRUE   1     0
#> 115 108 - 110  TRUE   1     0
#> 116 110 - 112  TRUE   1     0
#> 117 112 - 114  TRUE   1     0
#> 118 114 - 116  TRUE   1     0
#> 119 116 - 118  TRUE   1     0
#> 120 118 - 120  TRUE   1     0
#> 121     0 - 2 FALSE   2   217
#> 122     2 - 4 FALSE   2   240
#> 123     4 - 6 FALSE   2   228
#> 124     6 - 8 FALSE   2   205
#> 125    8 - 10 FALSE   2   216
#> 126   10 - 12 FALSE   2   240
#> 127   12 - 14 FALSE   2   184
#> 128   14 - 16 FALSE   2   204
#> 129   16 - 18 FALSE   2   226
#> 130   18 - 20 FALSE   2   218
#> 131   20 - 22 FALSE   2   186
#> 132   22 - 24 FALSE   2   247
#> 133   24 - 26 FALSE   2   217
#> 134   26 - 28 FALSE   2   217
#> 135   28 - 30 FALSE   2   223
#> 136   30 - 32 FALSE   2   252
#> 137   32 - 34 FALSE   2   274
#> 138   34 - 36 FALSE   2   248
#> 139   36 - 38 FALSE   2   227
#> 140   38 - 40 FALSE   2   248
#> 141   40 - 42 FALSE   2   266
#> 142   42 - 44 FALSE   2   279
#> 143   44 - 46 FALSE   2   294
#> 144   46 - 48 FALSE   2   293
#> 145   48 - 50 FALSE   2   324
#> 146   50 - 52 FALSE   2   321
#> 147   52 - 54 FALSE   2   324
#> 148   54 - 56 FALSE   2   284
#> 149   56 - 58 FALSE   2   244
#> 150   58 - 60 FALSE   2   238
#> 151   60 - 62 FALSE   2   264
#> 152   62 - 64 FALSE   2   243
#> 153   64 - 66 FALSE   2   274
#> 154   66 - 68 FALSE   2   278
#> 155   68 - 70 FALSE   2   221
#> 156   70 - 72 FALSE   2   241
#> 157   72 - 74 FALSE   2   161
#> 158   74 - 76 FALSE   2   150
#> 159   76 - 78 FALSE   2   161
#> 160   78 - 80 FALSE   2   141
#> 161   80 - 82 FALSE   2   133
#> 162   82 - 84 FALSE   2   119
#> 163   84 - 86 FALSE   2   107
#> 164   86 - 88 FALSE   2    97
#> 165   88 - 90 FALSE   2    75
#> 166   90 - 92 FALSE   2    48
#> 167   92 - 94 FALSE   2    38
#> 168   94 - 96 FALSE   2    19
#> 169   96 - 98 FALSE   2     4
#> 170  98 - 100 FALSE   2     4
#> 171 100 - 102 FALSE   2     2
#> 172 102 - 104 FALSE   2     0
#> 173 104 - 106 FALSE   2     0
#> 174 106 - 108 FALSE   2     0
#> 175 108 - 110 FALSE   2     0
#> 176 110 - 112 FALSE   2     0
#> 177 112 - 114 FALSE   2     0
#> 178 114 - 116 FALSE   2     0
#> 179 116 - 118 FALSE   2     0
#> 180 118 - 120 FALSE   2     0
#> 181     0 - 2  TRUE   2   208
#> 182     2 - 4  TRUE   2   240
#> 183     4 - 6  TRUE   2   250
#> 184     6 - 8  TRUE   2   202
#> 185    8 - 10  TRUE   2   234
#> 186   10 - 12  TRUE   2   207
#> 187   12 - 14  TRUE   2   204
#> 188   14 - 16  TRUE   2   226
#> 189   16 - 18  TRUE   2   234
#> 190   18 - 20  TRUE   2   241
#> 191   20 - 22  TRUE   2   266
#> 192   22 - 24  TRUE   2   254
#> 193   24 - 26  TRUE   2   241
#> 194   26 - 28  TRUE   2   231
#> 195   28 - 30  TRUE   2   241
#> 196   30 - 32  TRUE   2   221
#> 197   32 - 34  TRUE   2   252
#> 198   34 - 36  TRUE   2   245
#> 199   36 - 38  TRUE   2   221
#> 200   38 - 40  TRUE   2   219
#> 201   40 - 42  TRUE   2   242
#> 202   42 - 44  TRUE   2   311
#> 203   44 - 46  TRUE   2   278
#> 204   46 - 48  TRUE   2   293
#> 205   48 - 50  TRUE   2   326
#> 206   50 - 52  TRUE   2   282
#> 207   52 - 54  TRUE   2   273
#> 208   54 - 56  TRUE   2   261
#> 209   56 - 58  TRUE   2   258
#> 210   58 - 60  TRUE   2   241
#> 211   60 - 62  TRUE   2   225
#> 212   62 - 64  TRUE   2   242
#> 213   64 - 66  TRUE   2   236
#> 214   66 - 68  TRUE   2   250
#> 215   68 - 70  TRUE   2   197
#> 216   70 - 72  TRUE   2   177
#> 217   72 - 74  TRUE   2   145
#> 218   74 - 76  TRUE   2   157
#> 219   76 - 78  TRUE   2   142
#> 220   78 - 80  TRUE   2   127
#> 221   80 - 82  TRUE   2   118
#> 222   82 - 84  TRUE   2    89
#> 223   84 - 86  TRUE   2    75
#> 224   86 - 88  TRUE   2    55
#> 225   88 - 90  TRUE   2    48
#> 226   90 - 92  TRUE   2    29
#> 227   92 - 94  TRUE   2    15
#> 228   94 - 96  TRUE   2     3
#> 229   96 - 98  TRUE   2     6
#> 230  98 - 100  TRUE   2     2
#> 231 100 - 102  TRUE   2     0
#> 232 102 - 104  TRUE   2     0
#> 233 104 - 106  TRUE   2     0
#> 234 106 - 108  TRUE   2     0
#> 235 108 - 110  TRUE   2     1
#> 236 110 - 112  TRUE   2     0
#> 237 112 - 114  TRUE   2     0
#> 238 114 - 116  TRUE   2     0
#> 239 116 - 118  TRUE   2     0
#> 240 118 - 120  TRUE   2     0
#> 241     0 - 2 FALSE   3   220
#> 242     2 - 4 FALSE   3   200
#> 243     4 - 6 FALSE   3   243
#> 244     6 - 8 FALSE   3   237
#> 245    8 - 10 FALSE   3   213
#> 246   10 - 12 FALSE   3   176
#> 247   12 - 14 FALSE   3   216
#> 248   14 - 16 FALSE   3   228
#> 249   16 - 18 FALSE   3   238
#> 250   18 - 20 FALSE   3   212
#> 251   20 - 22 FALSE   3   257
#> 252   22 - 24 FALSE   3   287
#> 253   24 - 26 FALSE   3   263
#> 254   26 - 28 FALSE   3   252
#> 255   28 - 30 FALSE   3   267
#> 256   30 - 32 FALSE   3   255
#> 257   32 - 34 FALSE   3   254
#> 258   34 - 36 FALSE   3   324
#> 259   36 - 38 FALSE   3   243
#> 260   38 - 40 FALSE   3   243
#> 261   40 - 42 FALSE   3   255
#> 262   42 - 44 FALSE   3   296
#> 263   44 - 46 FALSE   3   252
#> 264   46 - 48 FALSE   3   275
#> 265   48 - 50 FALSE   3   324
#> 266   50 - 52 FALSE   3   297
#> 267   52 - 54 FALSE   3   281
#> 268   54 - 56 FALSE   3   256
#> 269   56 - 58 FALSE   3   240
#> 270   58 - 60 FALSE   3   239
#> 271   60 - 62 FALSE   3   262
#> 272   62 - 64 FALSE   3   249
#> 273   64 - 66 FALSE   3   231
#> 274   66 - 68 FALSE   3   290
#> 275   68 - 70 FALSE   3   207
#> 276   70 - 72 FALSE   3   177
#> 277   72 - 74 FALSE   3   176
#> 278   74 - 76 FALSE   3   171
#> 279   76 - 78 FALSE   3   151
#> 280   78 - 80 FALSE   3   151
#> 281   80 - 82 FALSE   3   138
#> 282   82 - 84 FALSE   3   112
#> 283   84 - 86 FALSE   3   108
#> 284   86 - 88 FALSE   3    78
#> 285   88 - 90 FALSE   3    74
#> 286   90 - 92 FALSE   3    67
#> 287   92 - 94 FALSE   3    37
#> 288   94 - 96 FALSE   3    23
#> 289   96 - 98 FALSE   3    14
#> 290  98 - 100 FALSE   3    11
#> 291 100 - 102 FALSE   3     3
#> 292 102 - 104 FALSE   3     3
#> 293 104 - 106 FALSE   3     1
#> 294 106 - 108 FALSE   3     0
#> 295 108 - 110 FALSE   3     0
#> 296 110 - 112 FALSE   3     0
#> 297 112 - 114 FALSE   3     0
#> 298 114 - 116 FALSE   3     0
#> 299 116 - 118 FALSE   3     0
#> 300 118 - 120 FALSE   3     0
#> 301     0 - 2  TRUE   3   219
#> 302     2 - 4  TRUE   3   255
#> 303     4 - 6  TRUE   3   207
#> 304     6 - 8  TRUE   3   258
#> 305    8 - 10  TRUE   3   214
#> 306   10 - 12  TRUE   3   187
#> 307   12 - 14  TRUE   3   205
#> 308   14 - 16  TRUE   3   193
#> 309   16 - 18  TRUE   3   256
#> 310   18 - 20  TRUE   3   234
#> 311   20 - 22  TRUE   3   246
#> 312   22 - 24  TRUE   3   286
#> 313   24 - 26  TRUE   3   261
#> 314   26 - 28  TRUE   3   287
#> 315   28 - 30  TRUE   3   268
#> 316   30 - 32  TRUE   3   275
#> 317   32 - 34  TRUE   3   297
#> 318   34 - 36  TRUE   3   286
#> 319   36 - 38  TRUE   3   232
#> 320   38 - 40  TRUE   3   245
#> 321   40 - 42  TRUE   3   267
#> 322   42 - 44  TRUE   3   263
#> 323   44 - 46  TRUE   3   277
#> 324   46 - 48  TRUE   3   260
#> 325   48 - 50  TRUE   3   297
#> 326   50 - 52  TRUE   3   270
#> 327   52 - 54  TRUE   3   240
#> 328   54 - 56  TRUE   3   222
#> 329   56 - 58  TRUE   3   259
#> 330   58 - 60  TRUE   3   223
#> 331   60 - 62  TRUE   3   222
#> 332   62 - 64  TRUE   3   238
#> 333   64 - 66  TRUE   3   226
#> 334   66 - 68  TRUE   3   270
#> 335   68 - 70  TRUE   3   197
#> 336   70 - 72  TRUE   3   180
#> 337   72 - 74  TRUE   3   171
#> 338   74 - 76  TRUE   3   150
#> 339   76 - 78  TRUE   3   163
#> 340   78 - 80  TRUE   3    90
#> 341   80 - 82  TRUE   3   113
#> 342   82 - 84  TRUE   3    72
#> 343   84 - 86  TRUE   3    70
#> 344   86 - 88  TRUE   3    52
#> 345   88 - 90  TRUE   3    33
#> 346   90 - 92  TRUE   3    22
#> 347   92 - 94  TRUE   3    17
#> 348   94 - 96  TRUE   3     8
#> 349   96 - 98  TRUE   3     2
#> 350  98 - 100  TRUE   3     2
#> 351 100 - 102  TRUE   3     0
#> 352 102 - 104  TRUE   3     0
#> 353 104 - 106  TRUE   3     0
#> 354 106 - 108  TRUE   3     0
#> 355 108 - 110  TRUE   3     0
#> 356 110 - 112  TRUE   3     0
#> 357 112 - 114  TRUE   3     0
#> 358 114 - 116  TRUE   3     0
#> 359 116 - 118  TRUE   3     0
#> 360 118 - 120  TRUE   3     0
#> 361     0 - 2 FALSE   4   262
#> 362     2 - 4 FALSE   4   300
#> 363     4 - 6 FALSE   4   263
#> 364     6 - 8 FALSE   4   264
#> 365    8 - 10 FALSE   4   213
#> 366   10 - 12 FALSE   4   217
#> 367   12 - 14 FALSE   4   204
#> 368   14 - 16 FALSE   4   213
#> 369   16 - 18 FALSE   4   237
#> 370   18 - 20 FALSE   4   262
#> 371   20 - 22 FALSE   4   276
#> 372   22 - 24 FALSE   4   313
#> 373   24 - 26 FALSE   4   354
#> 374   26 - 28 FALSE   4   331
#> 375   28 - 30 FALSE   4   315
#> 376   30 - 32 FALSE   4   320
#> 377   32 - 34 FALSE   4   341
#> 378   34 - 36 FALSE   4   304
#> 379   36 - 38 FALSE   4   280
#> 380   38 - 40 FALSE   4   258
#> 381   40 - 42 FALSE   4   249
#> 382   42 - 44 FALSE   4   266
#> 383   44 - 46 FALSE   4   289
#> 384   46 - 48 FALSE   4   271
#> 385   48 - 50 FALSE   4   255
#> 386   50 - 52 FALSE   4   270
#> 387   52 - 54 FALSE   4   292
#> 388   54 - 56 FALSE   4   250
#> 389   56 - 58 FALSE   4   250
#> 390   58 - 60 FALSE   4   222
#> 391   60 - 62 FALSE   4   194
#> 392   62 - 64 FALSE   4   214
#> 393   64 - 66 FALSE   4   182
#> 394   66 - 68 FALSE   4   209
#> 395   68 - 70 FALSE   4   178
#> 396   70 - 72 FALSE   4   172
#> 397   72 - 74 FALSE   4   138
#> 398   74 - 76 FALSE   4   145
#> 399   76 - 78 FALSE   4   133
#> 400   78 - 80 FALSE   4   121
#> 401   80 - 82 FALSE   4   106
#> 402   82 - 84 FALSE   4    95
#> 403   84 - 86 FALSE   4    85
#> 404   86 - 88 FALSE   4    64
#> 405   88 - 90 FALSE   4    60
#> 406   90 - 92 FALSE   4    45
#> 407   92 - 94 FALSE   4    39
#> 408   94 - 96 FALSE   4    11
#> 409   96 - 98 FALSE   4     8
#> 410  98 - 100 FALSE   4     4
#> 411 100 - 102 FALSE   4     2
#> 412 102 - 104 FALSE   4     0
#> 413 104 - 106 FALSE   4     0
#> 414 106 - 108 FALSE   4     0
#> 415 108 - 110 FALSE   4     1
#> 416 110 - 112 FALSE   4     0
#> 417 112 - 114 FALSE   4     0
#> 418 114 - 116 FALSE   4     0
#> 419 116 - 118 FALSE   4     0
#> 420 118 - 120 FALSE   4     0
#> 421     0 - 2  TRUE   4   263
#> 422     2 - 4  TRUE   4   313
#> 423     4 - 6  TRUE   4   267
#> 424     6 - 8  TRUE   4   259
#> 425    8 - 10  TRUE   4   261
#> 426   10 - 12  TRUE   4   219
#> 427   12 - 14  TRUE   4   213
#> 428   14 - 16  TRUE   4   239
#> 429   16 - 18  TRUE   4   224
#> 430   18 - 20  TRUE   4   251
#> 431   20 - 22  TRUE   4   320
#> 432   22 - 24  TRUE   4   340
#> 433   24 - 26  TRUE   4   356
#> 434   26 - 28  TRUE   4   317
#> 435   28 - 30  TRUE   4   350
#> 436   30 - 32  TRUE   4   357
#> 437   32 - 34  TRUE   4   341
#> 438   34 - 36  TRUE   4   323
#> 439   36 - 38  TRUE   4   286
#> 440   38 - 40  TRUE   4   271
#> 441   40 - 42  TRUE   4   278
#> 442   42 - 44  TRUE   4   304
#> 443   44 - 46  TRUE   4   289
#> 444   46 - 48  TRUE   4   255
#> 445   48 - 50  TRUE   4   267
#> 446   50 - 52  TRUE   4   282
#> 447   52 - 54  TRUE   4   250
#> 448   54 - 56  TRUE   4   236
#> 449   56 - 58  TRUE   4   215
#> 450   58 - 60  TRUE   4   192
#> 451   60 - 62  TRUE   4   195
#> 452   62 - 64  TRUE   4   200
#> 453   64 - 66  TRUE   4   205
#> 454   66 - 68  TRUE   4   193
#> 455   68 - 70  TRUE   4   159
#> 456   70 - 72  TRUE   4   146
#> 457   72 - 74  TRUE   4   118
#> 458   74 - 76  TRUE   4   114
#> 459   76 - 78  TRUE   4   119
#> 460   78 - 80  TRUE   4    94
#> 461   80 - 82  TRUE   4    72
#> 462   82 - 84  TRUE   4    75
#> 463   84 - 86  TRUE   4    65
#> 464   86 - 88  TRUE   4    44
#> 465   88 - 90  TRUE   4    33
#> 466   90 - 92  TRUE   4    24
#> 467   92 - 94  TRUE   4    19
#> 468   94 - 96  TRUE   4     8
#> 469   96 - 98  TRUE   4     6
#> 470  98 - 100  TRUE   4     1
#> 471 100 - 102  TRUE   4     0
#> 472 102 - 104  TRUE   4     0
#> 473 104 - 106  TRUE   4     0
#> 474 106 - 108  TRUE   4     0
#> 475 108 - 110  TRUE   4     0
#> 476 110 - 112  TRUE   4     0
#> 477 112 - 114  TRUE   4     0
#> 478 114 - 116  TRUE   4     0
#> 479 116 - 118  TRUE   4     0
#> 480 118 - 120  TRUE   4     0
#> 481     0 - 2 FALSE   5   308
#> 482     2 - 4 FALSE   5   353
#> 483     4 - 6 FALSE   5   296
#> 484     6 - 8 FALSE   5   294
#> 485    8 - 10 FALSE   5   264
#> 486   10 - 12 FALSE   5   243
#> 487   12 - 14 FALSE   5   239
#> 488   14 - 16 FALSE   5   259
#> 489   16 - 18 FALSE   5   243
#> 490   18 - 20 FALSE   5   290
#> 491   20 - 22 FALSE   5   276
#> 492   22 - 24 FALSE   5   339
#> 493   24 - 26 FALSE   5   345
#> 494   26 - 28 FALSE   5   354
#> 495   28 - 30 FALSE   5   376
#> 496   30 - 32 FALSE   5   309
#> 497   32 - 34 FALSE   5   304
#> 498   34 - 36 FALSE   5   303
#> 499   36 - 38 FALSE   5   285
#> 500   38 - 40 FALSE   5   246
#> 501   40 - 42 FALSE   5   230
#> 502   42 - 44 FALSE   5   255
#> 503   44 - 46 FALSE   5   259
#> 504   46 - 48 FALSE   5   268
#> 505   48 - 50 FALSE   5   232
#> 506   50 - 52 FALSE   5   241
#> 507   52 - 54 FALSE   5   233
#> 508   54 - 56 FALSE   5   231
#> 509   56 - 58 FALSE   5   194
#> 510   58 - 60 FALSE   5   203
#> 511   60 - 62 FALSE   5   183
#> 512   62 - 64 FALSE   5   153
#> 513   64 - 66 FALSE   5   171
#> 514   66 - 68 FALSE   5   166
#> 515   68 - 70 FALSE   5   130
#> 516   70 - 72 FALSE   5   135
#> 517   72 - 74 FALSE   5   131
#> 518   74 - 76 FALSE   5   116
#> 519   76 - 78 FALSE   5   119
#> 520   78 - 80 FALSE   5    98
#> 521   80 - 82 FALSE   5    87
#> 522   82 - 84 FALSE   5    89
#> 523   84 - 86 FALSE   5    67
#> 524   86 - 88 FALSE   5    59
#> 525   88 - 90 FALSE   5    55
#> 526   90 - 92 FALSE   5    40
#> 527   92 - 94 FALSE   5    26
#> 528   94 - 96 FALSE   5     5
#> 529   96 - 98 FALSE   5     3
#> 530  98 - 100 FALSE   5     5
#> 531 100 - 102 FALSE   5     1
#> 532 102 - 104 FALSE   5     1
#> 533 104 - 106 FALSE   5     0
#> 534 106 - 108 FALSE   5     0
#> 535 108 - 110 FALSE   5     0
#> 536 110 - 112 FALSE   5     0
#> 537 112 - 114 FALSE   5     0
#> 538 114 - 116 FALSE   5     0
#> 539 116 - 118 FALSE   5     0
#> 540 118 - 120 FALSE   5     0
#> 541     0 - 2  TRUE   5   333
#> 542     2 - 4  TRUE   5   325
#> 543     4 - 6  TRUE   5   318
#> 544     6 - 8  TRUE   5   324
#> 545    8 - 10  TRUE   5   269
#> 546   10 - 12  TRUE   5   240
#> 547   12 - 14  TRUE   5   239
#> 548   14 - 16  TRUE   5   273
#> 549   16 - 18  TRUE   5   254
#> 550   18 - 20  TRUE   5   296
#> 551   20 - 22  TRUE   5   281
#> 552   22 - 24  TRUE   5   304
#> 553   24 - 26  TRUE   5   324
#> 554   26 - 28  TRUE   5   326
#> 555   28 - 30  TRUE   5   350
#> 556   30 - 32  TRUE   5   292
#> 557   32 - 34  TRUE   5   315
#> 558   34 - 36  TRUE   5   296
#> 559   36 - 38  TRUE   5   249
#> 560   38 - 40  TRUE   5   291
#> 561   40 - 42  TRUE   5   241
#> 562   42 - 44  TRUE   5   275
#> 563   44 - 46  TRUE   5   320
#> 564   46 - 48  TRUE   5   267
#> 565   48 - 50  TRUE   5   261
#> 566   50 - 52  TRUE   5   256
#> 567   52 - 54  TRUE   5   229
#> 568   54 - 56  TRUE   5   222
#> 569   56 - 58  TRUE   5   205
#> 570   58 - 60  TRUE   5   189
#> 571   60 - 62  TRUE   5   178
#> 572   62 - 64  TRUE   5   173
#> 573   64 - 66  TRUE   5   151
#> 574   66 - 68  TRUE   5   176
#> 575   68 - 70  TRUE   5   141
#> 576   70 - 72  TRUE   5   123
#> 577   72 - 74  TRUE   5   110
#> 578   74 - 76  TRUE   5    93
#> 579   76 - 78  TRUE   5    90
#> 580   78 - 80  TRUE   5    71
#> 581   80 - 82  TRUE   5    78
#> 582   82 - 84  TRUE   5    49
#> 583   84 - 86  TRUE   5    53
#> 584   86 - 88  TRUE   5    40
#> 585   88 - 90  TRUE   5    14
#> 586   90 - 92  TRUE   5    12
#> 587   92 - 94  TRUE   5    13
#> 588   94 - 96  TRUE   5     4
#> 589   96 - 98  TRUE   5     1
#> 590  98 - 100  TRUE   5     1
#> 591 100 - 102  TRUE   5     0
#> 592 102 - 104  TRUE   5     1
#> 593 104 - 106  TRUE   5     0
#> 594 106 - 108  TRUE   5     0
#> 595 108 - 110  TRUE   5     0
#> 596 110 - 112  TRUE   5     0
#> 597 112 - 114  TRUE   5     0
#> 598 114 - 116  TRUE   5     0
#> 599 116 - 118  TRUE   5     0
#> 600 118 - 120  TRUE   5     0