Exposure table

Returns the Central Exposure-to-Risk for given ages groups and time period. The central Exposure-to-risk is computed as the sum of the time spent by individuals in a given age group over a given period, where age is the age at last birthday.

exposure_table(pop_df, ages, period)

Arguments

pop_df

Population data frame containing at least 'birth' and 'death' columns.

ages

A vector of size N composed of age groups.

period

A vector of size M composed of time intervals.

Value

An exposure matrix

Details

The function computes the central exposure-to-risk in each time interval [t[j],t[j+1]), j=1..M, and age groups.

Examples

ex_table <- exposure_table(EW_pop_out,0:101,0:2)