A function returning a merged dataframe from a list of population dataframes with id.

A function returning a merged dataframe from a list of population dataframes with id.

merge_pop_withid(pop_df_list, char_ignored = NULL)

Arguments

pop_df_list
A list of population dataframe where the first three columns of each dataframe are id, birth and death.

char_ignored
A vector of characteristics names which are only kept from the firt element of pop_df_list.

Value

A dataframe composed of all individuals with their characteristics at each simulation time.