Weibull function.

The Weibull (density) function is defined as $$h(x) = \bigl(\frac{k}{\lambda}\bigr) {\bigl( \frac{x}{\lambda} \bigr)}^{k-1} e^{-(x/\lambda)^k}$$ with \(k, \lambda \in (0, +\infty)\).

weibull(k, lambda = 1)

Arguments

k
Shape parameter, a positive real number.

lambda
Scale parameter, a positive real number, defaults to 1.

Value

The Weibull density function dweibull with shape parameter k and scale parameter lambda, see dweibull.

Details

A C++ version of this function is available. See vignette('IBMPopSim_cpp') for more details.

See also