Age pyramid from a population data frame at a given time.

Reduce a population data frame containing all individuals (with some characteristics) to an age-groups data frame (preserving characteristics). The function computes the number of individuals at time in each age group [ages[i],ages[i+1][, for i in {1,...,N-1}.

age_pyramid(population, time = 0, ages = c(0:110, Inf))

Arguments

population
Population data frame: characteristics in columns and individuals in rows. At least two columns birth (dates of birth) and death (dates of death) are required.

time
The age pyramid is computed at instant time. Must be a numeric greater than or equal to 0.

ages
(Optional) A numeric vector of distinct positive values composing age groups. Must be in increasing order.

Value

Age pyramid of a population at instant time. A data frame with columns age and value. Optional characteristics are preserved.

Details

See also age_pyramids.

Examples

age_pyramid(EW_pop_14$sample, time = 0)

#>           age  male value
#> 1       0 - 1 FALSE   581
#> 2       0 - 1  TRUE   584
#> 3       1 - 2 FALSE   577
#> 4       1 - 2  TRUE   605
#> 5       2 - 3 FALSE   638
#> 6       2 - 3  TRUE   718
#> 7       3 - 4 FALSE   632
#> 8       3 - 4  TRUE   675
#> 9       4 - 5 FALSE   590
#> 10      4 - 5  TRUE   608
#> 11      5 - 6 FALSE   570
#> 12      5 - 6  TRUE   649
#> 13      6 - 7 FALSE   579
#> 14      6 - 7  TRUE   682
#> 15      7 - 8 FALSE   588
#> 16      7 - 8  TRUE   687
#> 17      8 - 9 FALSE   548
#> 18      8 - 9  TRUE   582
#> 19     9 - 10 FALSE   532
#> 20     9 - 10  TRUE   570
#> 21    10 - 11 FALSE   503
#> 22    10 - 11  TRUE   584
#> 23    11 - 12 FALSE   562
#> 24    11 - 12  TRUE   600
#> 25    12 - 13 FALSE   514
#> 26    12 - 13  TRUE   518
#> 27    13 - 14 FALSE   509
#> 28    13 - 14  TRUE   589
#> 29    14 - 15 FALSE   573
#> 30    14 - 15  TRUE   606
#> 31    15 - 16 FALSE   582
#> 32    15 - 16  TRUE   609
#> 33    16 - 17 FALSE   536
#> 34    16 - 17  TRUE   607
#> 35    17 - 18 FALSE   565
#> 36    17 - 18  TRUE   612
#> 37    18 - 19 FALSE   618
#> 38    18 - 19  TRUE   624
#> 39    19 - 20 FALSE   616
#> 40    19 - 20  TRUE   657
#> 41    20 - 21 FALSE   623
#> 42    20 - 21  TRUE   656
#> 43    21 - 22 FALSE   650
#> 44    21 - 22  TRUE   614
#> 45    22 - 23 FALSE   714
#> 46    22 - 23  TRUE   684
#> 47    23 - 24 FALSE   671
#> 48    23 - 24  TRUE   673
#> 49    24 - 25 FALSE   621
#> 50    24 - 25  TRUE   716
#> 51    25 - 26 FALSE   682
#> 52    25 - 26  TRUE   669
#> 53    26 - 27 FALSE   714
#> 54    26 - 27  TRUE   723
#> 55    27 - 28 FALSE   687
#> 56    27 - 28  TRUE   625
#> 57    28 - 29 FALSE   712
#> 58    28 - 29  TRUE   665
#> 59    29 - 30 FALSE   723
#> 60    29 - 30  TRUE   729
#> 61    30 - 31 FALSE   696
#> 62    30 - 31  TRUE   671
#> 63    31 - 32 FALSE   670
#> 64    31 - 32  TRUE   663
#> 65    32 - 33 FALSE   725
#> 66    32 - 33  TRUE   650
#> 67    33 - 34 FALSE   695
#> 68    33 - 34  TRUE   698
#> 69    34 - 35 FALSE   692
#> 70    34 - 35  TRUE   647
#> 71    35 - 36 FALSE   657
#> 72    35 - 36  TRUE   665
#> 73    36 - 37 FALSE   609
#> 74    36 - 37  TRUE   590
#> 75    37 - 38 FALSE   608
#> 76    37 - 38  TRUE   607
#> 77    38 - 39 FALSE   644
#> 78    38 - 39  TRUE   618
#> 79    39 - 40 FALSE   600
#> 80    39 - 40  TRUE   612
#> 81    40 - 41 FALSE   612
#> 82    40 - 41  TRUE   632
#> 83    41 - 42 FALSE   642
#> 84    41 - 42  TRUE   655
#> 85    42 - 43 FALSE   757
#> 86    42 - 43  TRUE   681
#> 87    43 - 44 FALSE   728
#> 88    43 - 44  TRUE   743
#> 89    44 - 45 FALSE   685
#> 90    44 - 45  TRUE   701
#> 91    45 - 46 FALSE   754
#> 92    45 - 46  TRUE   689
#> 93    46 - 47 FALSE   729
#> 94    46 - 47  TRUE   718
#> 95    47 - 48 FALSE   728
#> 96    47 - 48  TRUE   729
#> 97    48 - 49 FALSE   752
#> 98    48 - 49  TRUE   666
#> 99    49 - 50 FALSE   773
#> 100   49 - 50  TRUE   724
#> 101   50 - 51 FALSE   753
#> 102   50 - 51  TRUE   708
#> 103   51 - 52 FALSE   660
#> 104   51 - 52  TRUE   686
#> 105   52 - 53 FALSE   714
#> 106   52 - 53  TRUE   713
#> 107   53 - 54 FALSE   668
#> 108   53 - 54  TRUE   610
#> 109   54 - 55 FALSE   616
#> 110   54 - 55  TRUE   648
#> 111   55 - 56 FALSE   626
#> 112   55 - 56  TRUE   560
#> 113   56 - 57 FALSE   680
#> 114   56 - 57  TRUE   622
#> 115   57 - 58 FALSE   609
#> 116   57 - 58  TRUE   612
#> 117   58 - 59 FALSE   558
#> 118   58 - 59  TRUE   552
#> 119   59 - 60 FALSE   571
#> 120   59 - 60  TRUE   550
#> 121   60 - 61 FALSE   547
#> 122   60 - 61  TRUE   575
#> 123   61 - 62 FALSE   571
#> 124   61 - 62  TRUE   558
#> 125   62 - 63 FALSE   534
#> 126   62 - 63  TRUE   540
#> 127   63 - 64 FALSE   521
#> 128   63 - 64  TRUE   502
#> 129   64 - 65 FALSE   533
#> 130   64 - 65  TRUE   487
#> 131   65 - 66 FALSE   580
#> 132   65 - 66  TRUE   543
#> 133   66 - 67 FALSE   640
#> 134   66 - 67  TRUE   588
#> 135   67 - 68 FALSE   643
#> 136   67 - 68  TRUE   636
#> 137   68 - 69 FALSE   504
#> 138   68 - 69  TRUE   468
#> 139   69 - 70 FALSE   467
#> 140   69 - 70  TRUE   477
#> 141   70 - 71 FALSE   501
#> 142   70 - 71  TRUE   390
#> 143   71 - 72 FALSE   417
#> 144   71 - 72  TRUE   455
#> 145   72 - 73 FALSE   434
#> 146   72 - 73  TRUE   371
#> 147   73 - 74 FALSE   364
#> 148   73 - 74  TRUE   347
#> 149   74 - 75 FALSE   404
#> 150   74 - 75  TRUE   338
#> 151   75 - 76 FALSE   370
#> 152   75 - 76  TRUE   312
#> 153   76 - 77 FALSE   338
#> 154   76 - 77  TRUE   327
#> 155   77 - 78 FALSE   380
#> 156   77 - 78  TRUE   300
#> 157   78 - 79 FALSE   343
#> 158   78 - 79  TRUE   282
#> 159   79 - 80 FALSE   341
#> 160   79 - 80  TRUE   261
#> 161   80 - 81 FALSE   295
#> 162   80 - 81  TRUE   218
#> 163   81 - 82 FALSE   260
#> 164   81 - 82  TRUE   218
#> 165   82 - 83 FALSE   275
#> 166   82 - 83  TRUE   228
#> 167   83 - 84 FALSE   259
#> 168   83 - 84  TRUE   177
#> 169   84 - 85 FALSE   231
#> 170   84 - 85  TRUE   160
#> 171   85 - 86 FALSE   239
#> 172   85 - 86  TRUE   132
#> 173   86 - 87 FALSE   217
#> 174   86 - 87  TRUE   137
#> 175   87 - 88 FALSE   207
#> 176   87 - 88  TRUE   114
#> 177   88 - 89 FALSE   168
#> 178   88 - 89  TRUE   102
#> 179   89 - 90 FALSE   133
#> 180   89 - 90  TRUE    76
#> 181   90 - 91 FALSE   125
#> 182   90 - 91  TRUE    49
#> 183   91 - 92 FALSE   112
#> 184   91 - 92  TRUE    60
#> 185   92 - 93 FALSE    92
#> 186   92 - 93  TRUE    38
#> 187   93 - 94 FALSE    79
#> 188   93 - 94  TRUE    23
#> 189   94 - 95 FALSE    49
#> 190   94 - 95  TRUE    18
#> 191   95 - 96 FALSE    36
#> 192   95 - 96  TRUE    11
#> 193   96 - 97 FALSE    25
#> 194   96 - 97  TRUE     8
#> 195   97 - 98 FALSE    18
#> 196   97 - 98  TRUE     4
#> 197   98 - 99 FALSE    10
#> 198   98 - 99  TRUE     3
#> 199  99 - 100 FALSE    12
#> 200  99 - 100  TRUE     4
#> 201 100 - 101 FALSE     8
#> 202 100 - 101  TRUE     1
#> 203 101 - 102 FALSE     6
#> 204 101 - 102  TRUE     4
#> 205 102 - 103 FALSE     3
#> 206 102 - 103  TRUE     0
#> 207 103 - 104 FALSE     2
#> 208 103 - 104  TRUE     0
#> 209 104 - 105 FALSE     1
#> 210 104 - 105  TRUE     1
#> 211 105 - 106 FALSE     0
#> 212 105 - 106  TRUE     0
#> 213 106 - 107 FALSE     2
#> 214 106 - 107  TRUE     0
#> 215 107 - 108 FALSE     0
#> 216 107 - 108  TRUE     0
#> 217 108 - 109 FALSE     0
#> 218 108 - 109  TRUE     0
#> 219 109 - 110 FALSE     0
#> 220 109 - 110  TRUE     0
#> 221 110 - Inf FALSE     0
#> 222 110 - Inf  TRUE     0

age_pyramid(EW_popIMD_14$sample, time = 0, ages = seq(0, 120, by=2))

#>           age  male IMD value
#> 1       0 - 2 FALSE   1   166
#> 2       0 - 2 FALSE   2   217
#> 3       0 - 2 FALSE   3   220
#> 4       0 - 2 FALSE   4   262
#> 5       0 - 2 FALSE   5   308
#> 6       0 - 2  TRUE   1   203
#> 7       0 - 2  TRUE   2   208
#> 8       0 - 2  TRUE   3   219
#> 9       0 - 2  TRUE   4   263
#> 10      0 - 2  TRUE   5   333
#> 11      2 - 4 FALSE   1   203
#> 12      2 - 4 FALSE   2   240
#> 13      2 - 4 FALSE   3   200
#> 14      2 - 4 FALSE   4   300
#> 15      2 - 4 FALSE   5   353
#> 16      2 - 4  TRUE   1   243
#> 17      2 - 4  TRUE   2   240
#> 18      2 - 4  TRUE   3   255
#> 19      2 - 4  TRUE   4   313
#> 20      2 - 4  TRUE   5   325
#> 21      4 - 6 FALSE   1   219
#> 22      4 - 6 FALSE   2   228
#> 23      4 - 6 FALSE   3   243
#> 24      4 - 6 FALSE   4   263
#> 25      4 - 6 FALSE   5   296
#> 26      4 - 6  TRUE   1   219
#> 27      4 - 6  TRUE   2   250
#> 28      4 - 6  TRUE   3   207
#> 29      4 - 6  TRUE   4   267
#> 30      4 - 6  TRUE   5   318
#> 31      6 - 8 FALSE   1   255
#> 32      6 - 8 FALSE   2   205
#> 33      6 - 8 FALSE   3   237
#> 34      6 - 8 FALSE   4   264
#> 35      6 - 8 FALSE   5   294
#> 36      6 - 8  TRUE   1   256
#> 37      6 - 8  TRUE   2   202
#> 38      6 - 8  TRUE   3   258
#> 39      6 - 8  TRUE   4   259
#> 40      6 - 8  TRUE   5   324
#> 41     8 - 10 FALSE   1   203
#> 42     8 - 10 FALSE   2   216
#> 43     8 - 10 FALSE   3   213
#> 44     8 - 10 FALSE   4   213
#> 45     8 - 10 FALSE   5   264
#> 46     8 - 10  TRUE   1   208
#> 47     8 - 10  TRUE   2   234
#> 48     8 - 10  TRUE   3   214
#> 49     8 - 10  TRUE   4   261
#> 50     8 - 10  TRUE   5   269
#> 51    10 - 12 FALSE   1   192
#> 52    10 - 12 FALSE   2   240
#> 53    10 - 12 FALSE   3   176
#> 54    10 - 12 FALSE   4   217
#> 55    10 - 12 FALSE   5   243
#> 56    10 - 12  TRUE   1   212
#> 57    10 - 12  TRUE   2   207
#> 58    10 - 12  TRUE   3   187
#> 59    10 - 12  TRUE   4   219
#> 60    10 - 12  TRUE   5   240
#> 61    12 - 14 FALSE   1   235
#> 62    12 - 14 FALSE   2   184
#> 63    12 - 14 FALSE   3   216
#> 64    12 - 14 FALSE   4   204
#> 65    12 - 14 FALSE   5   239
#> 66    12 - 14  TRUE   1   205
#> 67    12 - 14  TRUE   2   204
#> 68    12 - 14  TRUE   3   205
#> 69    12 - 14  TRUE   4   213
#> 70    12 - 14  TRUE   5   239
#> 71    14 - 16 FALSE   1   239
#> 72    14 - 16 FALSE   2   204
#> 73    14 - 16 FALSE   3   228
#> 74    14 - 16 FALSE   4   213
#> 75    14 - 16 FALSE   5   259
#> 76    14 - 16  TRUE   1   237
#> 77    14 - 16  TRUE   2   226
#> 78    14 - 16  TRUE   3   193
#> 79    14 - 16  TRUE   4   239
#> 80    14 - 16  TRUE   5   273
#> 81    16 - 18 FALSE   1   250
#> 82    16 - 18 FALSE   2   226
#> 83    16 - 18 FALSE   3   238
#> 84    16 - 18 FALSE   4   237
#> 85    16 - 18 FALSE   5   243
#> 86    16 - 18  TRUE   1   259
#> 87    16 - 18  TRUE   2   234
#> 88    16 - 18  TRUE   3   256
#> 89    16 - 18  TRUE   4   224
#> 90    16 - 18  TRUE   5   254
#> 91    18 - 20 FALSE   1   213
#> 92    18 - 20 FALSE   2   218
#> 93    18 - 20 FALSE   3   212
#> 94    18 - 20 FALSE   4   262
#> 95    18 - 20 FALSE   5   290
#> 96    18 - 20  TRUE   1   213
#> 97    18 - 20  TRUE   2   241
#> 98    18 - 20  TRUE   3   234
#> 99    18 - 20  TRUE   4   251
#> 100   18 - 20  TRUE   5   296
#> 101   20 - 22 FALSE   1   186
#> 102   20 - 22 FALSE   2   186
#> 103   20 - 22 FALSE   3   257
#> 104   20 - 22 FALSE   4   276
#> 105   20 - 22 FALSE   5   276
#> 106   20 - 22  TRUE   1   224
#> 107   20 - 22  TRUE   2   266
#> 108   20 - 22  TRUE   3   246
#> 109   20 - 22  TRUE   4   320
#> 110   20 - 22  TRUE   5   281
#> 111   22 - 24 FALSE   1   193
#> 112   22 - 24 FALSE   2   247
#> 113   22 - 24 FALSE   3   287
#> 114   22 - 24 FALSE   4   313
#> 115   22 - 24 FALSE   5   339
#> 116   22 - 24  TRUE   1   205
#> 117   22 - 24  TRUE   2   254
#> 118   22 - 24  TRUE   3   286
#> 119   22 - 24  TRUE   4   340
#> 120   22 - 24  TRUE   5   304
#> 121   24 - 26 FALSE   1   200
#> 122   24 - 26 FALSE   2   217
#> 123   24 - 26 FALSE   3   263
#> 124   24 - 26 FALSE   4   354
#> 125   24 - 26 FALSE   5   345
#> 126   24 - 26  TRUE   1   202
#> 127   24 - 26  TRUE   2   241
#> 128   24 - 26  TRUE   3   261
#> 129   24 - 26  TRUE   4   356
#> 130   24 - 26  TRUE   5   324
#> 131   26 - 28 FALSE   1   171
#> 132   26 - 28 FALSE   2   217
#> 133   26 - 28 FALSE   3   252
#> 134   26 - 28 FALSE   4   331
#> 135   26 - 28 FALSE   5   354
#> 136   26 - 28  TRUE   1   183
#> 137   26 - 28  TRUE   2   231
#> 138   26 - 28  TRUE   3   287
#> 139   26 - 28  TRUE   4   317
#> 140   26 - 28  TRUE   5   326
#> 141   28 - 30 FALSE   1   157
#> 142   28 - 30 FALSE   2   223
#> 143   28 - 30 FALSE   3   267
#> 144   28 - 30 FALSE   4   315
#> 145   28 - 30 FALSE   5   376
#> 146   28 - 30  TRUE   1   171
#> 147   28 - 30  TRUE   2   241
#> 148   28 - 30  TRUE   3   268
#> 149   28 - 30  TRUE   4   350
#> 150   28 - 30  TRUE   5   350
#> 151   30 - 32 FALSE   1   198
#> 152   30 - 32 FALSE   2   252
#> 153   30 - 32 FALSE   3   255
#> 154   30 - 32 FALSE   4   320
#> 155   30 - 32 FALSE   5   309
#> 156   30 - 32  TRUE   1   177
#> 157   30 - 32  TRUE   2   221
#> 158   30 - 32  TRUE   3   275
#> 159   30 - 32  TRUE   4   357
#> 160   30 - 32  TRUE   5   292
#> 161   32 - 34 FALSE   1   227
#> 162   32 - 34 FALSE   2   274
#> 163   32 - 34 FALSE   3   254
#> 164   32 - 34 FALSE   4   341
#> 165   32 - 34 FALSE   5   304
#> 166   32 - 34  TRUE   1   221
#> 167   32 - 34  TRUE   2   252
#> 168   32 - 34  TRUE   3   297
#> 169   32 - 34  TRUE   4   341
#> 170   32 - 34  TRUE   5   315
#> 171   34 - 36 FALSE   1   255
#> 172   34 - 36 FALSE   2   248
#> 173   34 - 36 FALSE   3   324
#> 174   34 - 36 FALSE   4   304
#> 175   34 - 36 FALSE   5   303
#> 176   34 - 36  TRUE   1   183
#> 177   34 - 36  TRUE   2   245
#> 178   34 - 36  TRUE   3   286
#> 179   34 - 36  TRUE   4   323
#> 180   34 - 36  TRUE   5   296
#> 181   36 - 38 FALSE   1   216
#> 182   36 - 38 FALSE   2   227
#> 183   36 - 38 FALSE   3   243
#> 184   36 - 38 FALSE   4   280
#> 185   36 - 38 FALSE   5   285
#> 186   36 - 38  TRUE   1   212
#> 187   36 - 38  TRUE   2   221
#> 188   36 - 38  TRUE   3   232
#> 189   36 - 38  TRUE   4   286
#> 190   36 - 38  TRUE   5   249
#> 191   38 - 40 FALSE   1   238
#> 192   38 - 40 FALSE   2   248
#> 193   38 - 40 FALSE   3   243
#> 194   38 - 40 FALSE   4   258
#> 195   38 - 40 FALSE   5   246
#> 196   38 - 40  TRUE   1   224
#> 197   38 - 40  TRUE   2   219
#> 198   38 - 40  TRUE   3   245
#> 199   38 - 40  TRUE   4   271
#> 200   38 - 40  TRUE   5   291
#> 201   40 - 42 FALSE   1   272
#> 202   40 - 42 FALSE   2   266
#> 203   40 - 42 FALSE   3   255
#> 204   40 - 42 FALSE   4   249
#> 205   40 - 42 FALSE   5   230
#> 206   40 - 42  TRUE   1   255
#> 207   40 - 42  TRUE   2   242
#> 208   40 - 42  TRUE   3   267
#> 209   40 - 42  TRUE   4   278
#> 210   40 - 42  TRUE   5   241
#> 211   42 - 44 FALSE   1   304
#> 212   42 - 44 FALSE   2   279
#> 213   42 - 44 FALSE   3   296
#> 214   42 - 44 FALSE   4   266
#> 215   42 - 44 FALSE   5   255
#> 216   42 - 44  TRUE   1   255
#> 217   42 - 44  TRUE   2   311
#> 218   42 - 44  TRUE   3   263
#> 219   42 - 44  TRUE   4   304
#> 220   42 - 44  TRUE   5   275
#> 221   44 - 46 FALSE   1   315
#> 222   44 - 46 FALSE   2   294
#> 223   44 - 46 FALSE   3   252
#> 224   44 - 46 FALSE   4   289
#> 225   44 - 46 FALSE   5   259
#> 226   44 - 46  TRUE   1   278
#> 227   44 - 46  TRUE   2   278
#> 228   44 - 46  TRUE   3   277
#> 229   44 - 46  TRUE   4   289
#> 230   44 - 46  TRUE   5   320
#> 231   46 - 48 FALSE   1   304
#> 232   46 - 48 FALSE   2   293
#> 233   46 - 48 FALSE   3   275
#> 234   46 - 48 FALSE   4   271
#> 235   46 - 48 FALSE   5   268
#> 236   46 - 48  TRUE   1   315
#> 237   46 - 48  TRUE   2   293
#> 238   46 - 48  TRUE   3   260
#> 239   46 - 48  TRUE   4   255
#> 240   46 - 48  TRUE   5   267
#> 241   48 - 50 FALSE   1   307
#> 242   48 - 50 FALSE   2   324
#> 243   48 - 50 FALSE   3   324
#> 244   48 - 50 FALSE   4   255
#> 245   48 - 50 FALSE   5   232
#> 246   48 - 50  TRUE   1   293
#> 247   48 - 50  TRUE   2   326
#> 248   48 - 50  TRUE   3   297
#> 249   48 - 50  TRUE   4   267
#> 250   48 - 50  TRUE   5   261
#> 251   50 - 52 FALSE   1   338
#> 252   50 - 52 FALSE   2   321
#> 253   50 - 52 FALSE   3   297
#> 254   50 - 52 FALSE   4   270
#> 255   50 - 52 FALSE   5   241
#> 256   50 - 52  TRUE   1   302
#> 257   50 - 52  TRUE   2   282
#> 258   50 - 52  TRUE   3   270
#> 259   50 - 52  TRUE   4   282
#> 260   50 - 52  TRUE   5   256
#> 261   52 - 54 FALSE   1   289
#> 262   52 - 54 FALSE   2   324
#> 263   52 - 54 FALSE   3   281
#> 264   52 - 54 FALSE   4   292
#> 265   52 - 54 FALSE   5   233
#> 266   52 - 54  TRUE   1   307
#> 267   52 - 54  TRUE   2   273
#> 268   52 - 54  TRUE   3   240
#> 269   52 - 54  TRUE   4   250
#> 270   52 - 54  TRUE   5   229
#> 271   54 - 56 FALSE   1   274
#> 272   54 - 56 FALSE   2   284
#> 273   54 - 56 FALSE   3   256
#> 274   54 - 56 FALSE   4   250
#> 275   54 - 56 FALSE   5   231
#> 276   54 - 56  TRUE   1   265
#> 277   54 - 56  TRUE   2   261
#> 278   54 - 56  TRUE   3   222
#> 279   54 - 56  TRUE   4   236
#> 280   54 - 56  TRUE   5   222
#> 281   56 - 58 FALSE   1   246
#> 282   56 - 58 FALSE   2   244
#> 283   56 - 58 FALSE   3   240
#> 284   56 - 58 FALSE   4   250
#> 285   56 - 58 FALSE   5   194
#> 286   56 - 58  TRUE   1   280
#> 287   56 - 58  TRUE   2   258
#> 288   56 - 58  TRUE   3   259
#> 289   56 - 58  TRUE   4   215
#> 290   56 - 58  TRUE   5   205
#> 291   58 - 60 FALSE   1   255
#> 292   58 - 60 FALSE   2   238
#> 293   58 - 60 FALSE   3   239
#> 294   58 - 60 FALSE   4   222
#> 295   58 - 60 FALSE   5   203
#> 296   58 - 60  TRUE   1   257
#> 297   58 - 60  TRUE   2   241
#> 298   58 - 60  TRUE   3   223
#> 299   58 - 60  TRUE   4   192
#> 300   58 - 60  TRUE   5   189
#> 301   60 - 62 FALSE   1   233
#> 302   60 - 62 FALSE   2   264
#> 303   60 - 62 FALSE   3   262
#> 304   60 - 62 FALSE   4   194
#> 305   60 - 62 FALSE   5   183
#> 306   60 - 62  TRUE   1   218
#> 307   60 - 62  TRUE   2   225
#> 308   60 - 62  TRUE   3   222
#> 309   60 - 62  TRUE   4   195
#> 310   60 - 62  TRUE   5   178
#> 311   62 - 64 FALSE   1   240
#> 312   62 - 64 FALSE   2   243
#> 313   62 - 64 FALSE   3   249
#> 314   62 - 64 FALSE   4   214
#> 315   62 - 64 FALSE   5   153
#> 316   62 - 64  TRUE   1   236
#> 317   62 - 64  TRUE   2   242
#> 318   62 - 64  TRUE   3   238
#> 319   62 - 64  TRUE   4   200
#> 320   62 - 64  TRUE   5   173
#> 321   64 - 66 FALSE   1   260
#> 322   64 - 66 FALSE   2   274
#> 323   64 - 66 FALSE   3   231
#> 324   64 - 66 FALSE   4   182
#> 325   64 - 66 FALSE   5   171
#> 326   64 - 66  TRUE   1   259
#> 327   64 - 66  TRUE   2   236
#> 328   64 - 66  TRUE   3   226
#> 329   64 - 66  TRUE   4   205
#> 330   64 - 66  TRUE   5   151
#> 331   66 - 68 FALSE   1   333
#> 332   66 - 68 FALSE   2   278
#> 333   66 - 68 FALSE   3   290
#> 334   66 - 68 FALSE   4   209
#> 335   66 - 68 FALSE   5   166
#> 336   66 - 68  TRUE   1   269
#> 337   66 - 68  TRUE   2   250
#> 338   66 - 68  TRUE   3   270
#> 339   66 - 68  TRUE   4   193
#> 340   66 - 68  TRUE   5   176
#> 341   68 - 70 FALSE   1   245
#> 342   68 - 70 FALSE   2   221
#> 343   68 - 70 FALSE   3   207
#> 344   68 - 70 FALSE   4   178
#> 345   68 - 70 FALSE   5   130
#> 346   68 - 70  TRUE   1   219
#> 347   68 - 70  TRUE   2   197
#> 348   68 - 70  TRUE   3   197
#> 349   68 - 70  TRUE   4   159
#> 350   68 - 70  TRUE   5   141
#> 351   70 - 72 FALSE   1   239
#> 352   70 - 72 FALSE   2   241
#> 353   70 - 72 FALSE   3   177
#> 354   70 - 72 FALSE   4   172
#> 355   70 - 72 FALSE   5   135
#> 356   70 - 72  TRUE   1   192
#> 357   70 - 72  TRUE   2   177
#> 358   70 - 72  TRUE   3   180
#> 359   70 - 72  TRUE   4   146
#> 360   70 - 72  TRUE   5   123
#> 361   72 - 74 FALSE   1   164
#> 362   72 - 74 FALSE   2   161
#> 363   72 - 74 FALSE   3   176
#> 364   72 - 74 FALSE   4   138
#> 365   72 - 74 FALSE   5   131
#> 366   72 - 74  TRUE   1   159
#> 367   72 - 74  TRUE   2   145
#> 368   72 - 74  TRUE   3   171
#> 369   72 - 74  TRUE   4   118
#> 370   72 - 74  TRUE   5   110
#> 371   74 - 76 FALSE   1   158
#> 372   74 - 76 FALSE   2   150
#> 373   74 - 76 FALSE   3   171
#> 374   74 - 76 FALSE   4   145
#> 375   74 - 76 FALSE   5   116
#> 376   74 - 76  TRUE   1   158
#> 377   74 - 76  TRUE   2   157
#> 378   74 - 76  TRUE   3   150
#> 379   74 - 76  TRUE   4   114
#> 380   74 - 76  TRUE   5    93
#> 381   76 - 78 FALSE   1   152
#> 382   76 - 78 FALSE   2   161
#> 383   76 - 78 FALSE   3   151
#> 384   76 - 78 FALSE   4   133
#> 385   76 - 78 FALSE   5   119
#> 386   76 - 78  TRUE   1   126
#> 387   76 - 78  TRUE   2   142
#> 388   76 - 78  TRUE   3   163
#> 389   76 - 78  TRUE   4   119
#> 390   76 - 78  TRUE   5    90
#> 391   78 - 80 FALSE   1   143
#> 392   78 - 80 FALSE   2   141
#> 393   78 - 80 FALSE   3   151
#> 394   78 - 80 FALSE   4   121
#> 395   78 - 80 FALSE   5    98
#> 396   78 - 80  TRUE   1   107
#> 397   78 - 80  TRUE   2   127
#> 398   78 - 80  TRUE   3    90
#> 399   78 - 80  TRUE   4    94
#> 400   78 - 80  TRUE   5    71
#> 401   80 - 82 FALSE   1   142
#> 402   80 - 82 FALSE   2   133
#> 403   80 - 82 FALSE   3   138
#> 404   80 - 82 FALSE   4   106
#> 405   80 - 82 FALSE   5    87
#> 406   80 - 82  TRUE   1   120
#> 407   80 - 82  TRUE   2   118
#> 408   80 - 82  TRUE   3   113
#> 409   80 - 82  TRUE   4    72
#> 410   80 - 82  TRUE   5    78
#> 411   82 - 84 FALSE   1   121
#> 412   82 - 84 FALSE   2   119
#> 413   82 - 84 FALSE   3   112
#> 414   82 - 84 FALSE   4    95
#> 415   82 - 84 FALSE   5    89
#> 416   82 - 84  TRUE   1   107
#> 417   82 - 84  TRUE   2    89
#> 418   82 - 84  TRUE   3    72
#> 419   82 - 84  TRUE   4    75
#> 420   82 - 84  TRUE   5    49
#> 421   84 - 86 FALSE   1   107
#> 422   84 - 86 FALSE   2   107
#> 423   84 - 86 FALSE   3   108
#> 424   84 - 86 FALSE   4    85
#> 425   84 - 86 FALSE   5    67
#> 426   84 - 86  TRUE   1    83
#> 427   84 - 86  TRUE   2    75
#> 428   84 - 86  TRUE   3    70
#> 429   84 - 86  TRUE   4    65
#> 430   84 - 86  TRUE   5    53
#> 431   86 - 88 FALSE   1    83
#> 432   86 - 88 FALSE   2    97
#> 433   86 - 88 FALSE   3    78
#> 434   86 - 88 FALSE   4    64
#> 435   86 - 88 FALSE   5    59
#> 436   86 - 88  TRUE   1    56
#> 437   86 - 88  TRUE   2    55
#> 438   86 - 88  TRUE   3    52
#> 439   86 - 88  TRUE   4    44
#> 440   86 - 88  TRUE   5    40
#> 441   88 - 90 FALSE   1    58
#> 442   88 - 90 FALSE   2    75
#> 443   88 - 90 FALSE   3    74
#> 444   88 - 90 FALSE   4    60
#> 445   88 - 90 FALSE   5    55
#> 446   88 - 90  TRUE   1    44
#> 447   88 - 90  TRUE   2    48
#> 448   88 - 90  TRUE   3    33
#> 449   88 - 90  TRUE   4    33
#> 450   88 - 90  TRUE   5    14
#> 451   90 - 92 FALSE   1    49
#> 452   90 - 92 FALSE   2    48
#> 453   90 - 92 FALSE   3    67
#> 454   90 - 92 FALSE   4    45
#> 455   90 - 92 FALSE   5    40
#> 456   90 - 92  TRUE   1    24
#> 457   90 - 92  TRUE   2    29
#> 458   90 - 92  TRUE   3    22
#> 459   90 - 92  TRUE   4    24
#> 460   90 - 92  TRUE   5    12
#> 461   92 - 94 FALSE   1    48
#> 462   92 - 94 FALSE   2    38
#> 463   92 - 94 FALSE   3    37
#> 464   92 - 94 FALSE   4    39
#> 465   92 - 94 FALSE   5    26
#> 466   92 - 94  TRUE   1    14
#> 467   92 - 94  TRUE   2    15
#> 468   92 - 94  TRUE   3    17
#> 469   92 - 94  TRUE   4    19
#> 470   92 - 94  TRUE   5    13
#> 471   94 - 96 FALSE   1    21
#> 472   94 - 96 FALSE   2    19
#> 473   94 - 96 FALSE   3    23
#> 474   94 - 96 FALSE   4    11
#> 475   94 - 96 FALSE   5     5
#> 476   94 - 96  TRUE   1    12
#> 477   94 - 96  TRUE   2     3
#> 478   94 - 96  TRUE   3     8
#> 479   94 - 96  TRUE   4     8
#> 480   94 - 96  TRUE   5     4
#> 481   96 - 98 FALSE   1     9
#> 482   96 - 98 FALSE   2     4
#> 483   96 - 98 FALSE   3    14
#> 484   96 - 98 FALSE   4     8
#> 485   96 - 98 FALSE   5     3
#> 486   96 - 98  TRUE   1     2
#> 487   96 - 98  TRUE   2     6
#> 488   96 - 98  TRUE   3     2
#> 489   96 - 98  TRUE   4     6
#> 490   96 - 98  TRUE   5     1
#> 491  98 - 100 FALSE   1     4
#> 492  98 - 100 FALSE   2     4
#> 493  98 - 100 FALSE   3    11
#> 494  98 - 100 FALSE   4     4
#> 495  98 - 100 FALSE   5     5
#> 496  98 - 100  TRUE   1     1
#> 497  98 - 100  TRUE   2     2
#> 498  98 - 100  TRUE   3     2
#> 499  98 - 100  TRUE   4     1
#> 500  98 - 100  TRUE   5     1
#> 501 100 - 102 FALSE   1     2
#> 502 100 - 102 FALSE   2     2
#> 503 100 - 102 FALSE   3     3
#> 504 100 - 102 FALSE   4     2
#> 505 100 - 102 FALSE   5     1
#> 506 100 - 102  TRUE   1     0
#> 507 100 - 102  TRUE   2     0
#> 508 100 - 102  TRUE   3     0
#> 509 100 - 102  TRUE   4     0
#> 510 100 - 102  TRUE   5     0
#> 511 102 - 104 FALSE   1     6
#> 512 102 - 104 FALSE   2     0
#> 513 102 - 104 FALSE   3     3
#> 514 102 - 104 FALSE   4     0
#> 515 102 - 104 FALSE   5     1
#> 516 102 - 104  TRUE   1     1
#> 517 102 - 104  TRUE   2     0
#> 518 102 - 104  TRUE   3     0
#> 519 102 - 104  TRUE   4     0
#> 520 102 - 104  TRUE   5     1
#> 521 104 - 106 FALSE   1     0
#> 522 104 - 106 FALSE   2     0
#> 523 104 - 106 FALSE   3     1
#> 524 104 - 106 FALSE   4     0
#> 525 104 - 106 FALSE   5     0
#> 526 104 - 106  TRUE   1     0
#> 527 104 - 106  TRUE   2     0
#> 528 104 - 106  TRUE   3     0
#> 529 104 - 106  TRUE   4     0
#> 530 104 - 106  TRUE   5     0
#> 531 106 - 108 FALSE   1     0
#> 532 106 - 108 FALSE   2     0
#> 533 106 - 108 FALSE   3     0
#> 534 106 - 108 FALSE   4     0
#> 535 106 - 108 FALSE   5     0
#> 536 106 - 108  TRUE   1     0
#> 537 106 - 108  TRUE   2     0
#> 538 106 - 108  TRUE   3     0
#> 539 106 - 108  TRUE   4     0
#> 540 106 - 108  TRUE   5     0
#> 541 108 - 110 FALSE   1     0
#> 542 108 - 110 FALSE   2     0
#> 543 108 - 110 FALSE   3     0
#> 544 108 - 110 FALSE   4     1
#> 545 108 - 110 FALSE   5     0
#> 546 108 - 110  TRUE   1     0
#> 547 108 - 110  TRUE   2     1
#> 548 108 - 110  TRUE   3     0
#> 549 108 - 110  TRUE   4     0
#> 550 108 - 110  TRUE   5     0
#> 551 110 - 112 FALSE   1     0
#> 552 110 - 112 FALSE   2     0
#> 553 110 - 112 FALSE   3     0
#> 554 110 - 112 FALSE   4     0
#> 555 110 - 112 FALSE   5     0
#> 556 110 - 112  TRUE   1     0
#> 557 110 - 112  TRUE   2     0
#> 558 110 - 112  TRUE   3     0
#> 559 110 - 112  TRUE   4     0
#> 560 110 - 112  TRUE   5     0
#> 561 112 - 114 FALSE   1     0
#> 562 112 - 114 FALSE   2     0
#> 563 112 - 114 FALSE   3     0
#> 564 112 - 114 FALSE   4     0
#> 565 112 - 114 FALSE   5     0
#> 566 112 - 114  TRUE   1     0
#> 567 112 - 114  TRUE   2     0
#> 568 112 - 114  TRUE   3     0
#> 569 112 - 114  TRUE   4     0
#> 570 112 - 114  TRUE   5     0
#> 571 114 - 116 FALSE   1     0
#> 572 114 - 116 FALSE   2     0
#> 573 114 - 116 FALSE   3     0
#> 574 114 - 116 FALSE   4     0
#> 575 114 - 116 FALSE   5     0
#> 576 114 - 116  TRUE   1     0
#> 577 114 - 116  TRUE   2     0
#> 578 114 - 116  TRUE   3     0
#> 579 114 - 116  TRUE   4     0
#> 580 114 - 116  TRUE   5     0
#> 581 116 - 118 FALSE   1     0
#> 582 116 - 118 FALSE   2     0
#> 583 116 - 118 FALSE   3     0
#> 584 116 - 118 FALSE   4     0
#> 585 116 - 118 FALSE   5     0
#> 586 116 - 118  TRUE   1     0
#> 587 116 - 118  TRUE   2     0
#> 588 116 - 118  TRUE   3     0
#> 589 116 - 118  TRUE   4     0
#> 590 116 - 118  TRUE   5     0
#> 591 118 - 120 FALSE   1     0
#> 592 118 - 120 FALSE   2     0
#> 593 118 - 120 FALSE   3     0
#> 594 118 - 120 FALSE   4     0
#> 595 118 - 120 FALSE   5     0
#> 596 118 - 120  TRUE   1     0
#> 597 118 - 120  TRUE   2     0
#> 598 118 - 120  TRUE   3     0
#> 599 118 - 120  TRUE   4     0
#> 600 118 - 120  TRUE   5     0