Creates Poisson class event.

The function mk_event_poisson is used to create an event with intensity of type Poisson (constant intensity which does not depend on population or time). When the event occurs, something happens in the population. The created event must be used with mk_model.

mk_event_poisson(type, name, intensity, kernel_code = "")

Arguments

type

Must be one of 'birth', 'death', 'entry', 'exit', 'swap' or 'custom'. See details.

name

(Optional) If not specified, the name given to the event is its type.

intensity

String containing some constant positive value, or name of a parameter which is a constant positive value.

kernel_code

String containing some C++ code describing the event action. Optional for 'birth', 'death' and 'exit' events. See details.

Value

An S3 object of class event of type Poisson.

Details

The type argument is one of the following

'birth'

By default, a new individual newI is created, with the same characteristics of the parent I and birth date equal to the current time. Optional code can be precised in kernel_code.

'death'

By default, the individual I dies. Optional code can be precised in kernel_code.

'entry'

A new individual newI is added to the population, and its characteristics have to be defined by the user in the entry kernel_code.

'exit'

An individual I exits from the population. Optional code can be precised in kernel_code.

'swap'

The user can change the characteristics of the selected individual I. This requires kernel_code.

'custom'

None of the above types, the user defines kernel_code that can act on the selected individual I and on the population pop.

The kernel_code argument is a string containing some C++ code which describing the action of the event. Some of available variables in the C++ code are: t (the current time), pop (the current population), I (the current individual selected for the event), newI (the new individual if 'birth' or 'entry' event), the name of the model parameters (some variables, or functions, see mk_model). See vignette('IBMPopSim') for more details.

See also

Examples

birth <- mk_event_poisson('birth', intensity = 10)

# \donttest{
params <- list(beta = 10)
death <- mk_event_poisson('death', intensity = 'beta') # name of one parameter
mk_model(events = list(birth, death), parameters = params)
#> $individual_type
#> $individual_type$names
#> [1] "birth" "death"
#> 
#> $individual_type$typesR
#> [1] "double" "double"
#> 
#> $individual_type$typesC
#> [1] "double" "double"
#> 
#> 
#> $parameters_types
#> $parameters_types$names
#> [1] "beta"
#> 
#> $parameters_types$typesR
#> [1] "double"
#> 
#> $parameters_types$typesC
#> [1] "double"
#> 
#> $parameters_types$lengths
#> beta 
#>    1 
#> 
#> 
#> $events
#> $events[[1]]
#> $name
#> [1] "birth"
#> 
#> $type
#> [1] "poisson" "birth"  
#> 
#> $intensity_code
#> [1] "10"
#> 
#> $kernel_code
#> [1] ""
#> 
#> $cpp_code
#> [1] "// Event \"birth\"\nstruct event_J_ : public event {\n    double __lambda(unsigned _k, population const & pop, double t, context & cntxt) const {\n        double result = cntxt.__intensity_bound[_J_];\n        return result;\n    }\n    double __lambda_bound(population const & pop, context & cntxt) const {\n        double result = cntxt.__intensity_bound[_J_] / cntxt.__num_threads;\n        return result;\n    }\n    void __kernel(unsigned _k, population & pop, double t, context & cntxt) const {\n        individual & I = pop[_k];\n        individual newI = I;\n\tnewI.birth_date = t;\n\t_CORRECTION_IF_ENTRY_\n\t\n\tpop.add(newI);\n    }\n    bool __apply(double t, population & pop, context & cntxt) const {\n        unsigned _k = pop.pick_alive(t, cntxt);\n        __kernel(_k, pop, t, cntxt);\n        return true;\n    }\n};\n"
#> 
#> attr(,"class")
#> [1] "event"
#> 
#> $events[[2]]
#> $name
#> [1] "death"
#> 
#> $type
#> [1] "poisson" "death"  
#> 
#> $intensity_code
#> [1] "beta"
#> 
#> $kernel_code
#> [1] ""
#> 
#> $cpp_code
#> [1] "// Event \"death\"\nstruct event_J_ : public event {\n    double __lambda(unsigned _k, population const & pop, double t, context & cntxt) const {\n        double result = cntxt.__intensity_bound[_J_];\n        return result;\n    }\n    double __lambda_bound(population const & pop, context & cntxt) const {\n        double result = cntxt.__intensity_bound[_J_] / cntxt.__num_threads;\n        return result;\n    }\n    void __kernel(unsigned _k, population & pop, double t, context & cntxt) const {\n        individual & I = pop[_k];\n        \n\tpop.kill(_k, t);\n    }\n    bool __apply(double t, population & pop, context & cntxt) const {\n        unsigned _k = pop.pick_alive(t, cntxt);\n        __kernel(_k, pop, t, cntxt);\n        return true;\n    }\n};\n"
#> 
#> attr(,"class")
#> [1] "event"
#> 
#> 
#> $cpp_code
#> [1] "// -*- coding: utf-8 -*-\n//[[Rcpp::depends(IBMPopSim)]]\n//[[Rcpp::depends(RcppArmadillo)]]\n//[[Rcpp::plugins(cpp11)]]\n#include <RcppArmadillo.h>\n#include <cmath>\n#include <limits>\n#include <array>\n#include <vector>\n#include <string>\n#include <thread>\n#include <chrono>\n\n#include \"functions.hpp\"\n\n// ***** DECLARATION_INDIVIDUAL\nstruct  individual{\n    individual() : death_date(NAN){}\n    individual(double birth_date,  double death_date = NAN)\n        : birth_date(birth_date),  death_date(death_date) { }\n    double birth_date;\n\n    // **** caracteristiques supplémentaires\n\n    // ****\n\n    double death_date; // If not dead, death_date = nan\n    double age(double t) const { return t - birth_date; };\n    void set_age(double a, double t) { birth_date = t-a; };\n    bool is_dead() const {return !std::isnan(death_date); };\n};\n\ninline double age(individual const & I, double t) { return I.age(t); }\n\n// *****\n\n// ***** DECLARATION_PARAMS MK_CONTEXT\nstruct context {\n    std::mt19937_64 __gen;\n    int __num_threads;\n    std::vector<double> __intensity_bound;\n\tdouble beta;\n#define beta cntxt.beta\n\n\n    template<typename ... Args>\n    double Unif(Args&&... arg) { return std::uniform_real_distribution<double>(std::forward<Args>(arg)...)(__gen); }\n#define CUnif cntxt.Unif\n    template<typename ... Args>\n    double Exp(Args&&... arg) { return std::exponential_distribution<double>(std::forward<Args>(arg)...)(__gen); }\n#define CExp  cntxt.Exp\n    template<typename ... Args>\n    double Norm(Args&&... arg) { return std::normal_distribution<double>(std::forward<Args>(arg)...)(__gen); }\n#define CNorm cntxt.Norm\n    template<typename ... Args>\n    unsigned Poisson(Args&&... arg) { return std::poisson_distribution<>(std::forward<Args>(arg)...)(__gen); }\n#define CPoisson cntxt.Poisson\n    template<typename ... Args>\n    double Gamma(Args&&... arg) { return std::gamma_distribution<double>(std::forward<Args>(arg)...)(__gen); }\n#define CGamma cntxt.Gamma\n    template<typename ... Args>\n    double Weibull(Args&&... arg) { return std::weibull_distribution<double>(std::forward<Args>(arg)...)(__gen); }\n#define CWeibull cntxt.Weibull\n    template<typename ... Args>\n    int UnifInt(Args&&... arg) { return std::uniform_int_distribution<int>(std::forward<Args>(arg)...)(__gen); }\n#define CUnifInt  cntxt.UnifInt\n    template<typename ... Args>\n    unsigned Discrete(Args&&... arg) { return std::discrete_distribution<int>(std::forward<Args>(arg)...)(__gen); }\n#define CDiscrete  cntxt.Discrete\n    template<typename ... Args>\n    bool Bernoulli(Args&&... arg) { return std::bernoulli_distribution(std::forward<Args>(arg)...)(__gen); }\n#define CBern  cntxt.Bernoulli\n    template<typename ... Args>\n    unsigned Binomial(Args&&... arg) { return std::binomial_distribution<int>(std::forward<Args>(arg)...)(__gen); }\n#define CBinom  cntxt.Binomial\n};\n// *****\n\n#include \"population.hpp\"\n\n// ***** DEFINITION_EVENTS\nstruct event {\n    virtual double __lambda(unsigned k, population const & pop, double t, context & cntxt) const = 0;\n    virtual double __lambda_bound(population const & pop, context & cntxt) const = 0;\n    virtual void __kernel(unsigned k, population & pop, double t, context & cntxt) const = 0;\n    virtual bool __apply(double t, population & pop, context & cntxt) const = 0;\n};\n\nconstexpr int NEVENTS = 2;\n// Event \"birth\"\nstruct event0 : public event {\n    double __lambda(unsigned _k, population const & pop, double t, context & cntxt) const {\n        double result = cntxt.__intensity_bound[0];\n        return result;\n    }\n    double __lambda_bound(population const & pop, context & cntxt) const {\n        double result = cntxt.__intensity_bound[0] / cntxt.__num_threads;\n        return result;\n    }\n    void __kernel(unsigned _k, population & pop, double t, context & cntxt) const {\n        individual & I = pop[_k];\n        individual newI = I;\n\tnewI.birth_date = t;\n\t\n\t\n\tpop.add(newI);\n    }\n    bool __apply(double t, population & pop, context & cntxt) const {\n        unsigned _k = pop.pick_alive(t, cntxt);\n        __kernel(_k, pop, t, cntxt);\n        return true;\n    }\n};\n// Event \"death\"\nstruct event1 : public event {\n    double __lambda(unsigned _k, population const & pop, double t, context & cntxt) const {\n        double result = cntxt.__intensity_bound[1];\n        return result;\n    }\n    double __lambda_bound(population const & pop, context & cntxt) const {\n        double result = cntxt.__intensity_bound[1] / cntxt.__num_threads;\n        return result;\n    }\n    void __kernel(unsigned _k, population & pop, double t, context & cntxt) const {\n        individual & I = pop[_k];\n        \n\tpop.kill(_k, t);\n    }\n    bool __apply(double t, population & pop, context & cntxt) const {\n        unsigned _k = pop.pick_alive(t, cntxt);\n        __kernel(_k, pop, t, cntxt);\n        return true;\n    }\n};\nstd::array<event*, NEVENTS> events { new event0, new event1 };\n// *****\n\n#include \"popsim_algorithm.hpp\"\n\nusing namespace Rcpp;\n\n//' @param pop_df Dataframe representing the initial population\n//' @param events_bounds NumericVector containing the events bounds\n//' @param parameters R parameters and functions available in C++ code\n//' @param age_max Age max of the population\n//' @param T0 Initial time\n//' @param T Final time\n//' @param clean_step Clean step for cleaning population from dead (or exit) individuals, by default equal to population age max\n//' @param clean_ratio Ratio of dead (or exit) individuals in the population for cleaning population from dead (or exit) individuals\n//' @param seed Seed for the random number generator\n//' @param multithreading Boolean for multithread activation, by default multithread is deactivated\n//' @param num_threads Number of threads for multithread run, if multihtread is activated the default value is the maximum of threads\n//' @param verbose verbose by default false\n//'\n//' @export\n// [[Rcpp::export]]\nRcpp::List popsim_cpp(Rcpp::DataFrame pop_df,\n        Rcpp::NumericVector events_bounds,\n        Rcpp::List parameters,\n        double age_max, double T0, double T,\n        double clean_step, double clean_ratio,\n        int seed,\n        bool multithreading, int num_threads = NA_INTEGER,\n        bool verbose = false)\n{\n    std::vector<double> intensity_bound = Rcpp::as<std::vector<double>>(events_bounds);\n    population pop(age_max);\n\n    if (verbose)\n        Rcpp::Rcout << \"Simulation on the interval [\" << T0 << \", \" << T << \"] \" << std::endl;\n    if (multithreading) {\n#ifdef __INTERACTION\n        Rcpp::Rcout << \"Warning: Multithreading should be deactivated...\" << std::endl;\n#endif\n        if (num_threads == NA_INTEGER) {\n            num_threads = std::thread::hardware_concurrency();\n        }\n        if (verbose) Rcpp::Rcout << \"Multithreading activated with \" << num_threads << \" threads\" << std::endl;\n    } else {\n        if (verbose) Rcpp::Rcout << \"Multithreading deactivated\" << std::endl;\n        num_threads = 1;\n    }\n\n    if (verbose) Rcpp::Rcout << \"Seed: \" << seed << std::endl;\n    std::vector<uint32_t> seeds(2*num_threads);\n    std::seed_seq seq { seed };\n    seq.generate(seeds.begin(), seeds.end());\n\n    // ***** INITIALISATION_POPULATION\n    Rcpp::DoubleVector _birth_date = pop_df[\"birth\"];\n\tRcpp::DoubleVector _death_date = pop_df[\"death\"];\n\tfor (unsigned _k = 0; _k < pop_df.nrow(); ++_k)\n\t    pop.add_init(individual(_birth_date[_k], _death_date[_k]));\n\t\n    // *****\n\n    // ***** DEFINITION_PARAMETERS\n    context cntxt {\n        std::mt19937_64(),\n        num_threads,\n        intensity_bound,\n        Rcpp::as<double>(parameters[\"beta\"])\n\t\t\n    };\n    // *****\n\n    std::vector<std::thread> threads(num_threads);\n    std::vector<counter> counters(num_threads);\n    std::vector<population> pops = split(pop, num_threads);\n\n    if (verbose) {\n        std::cout << \"before main algorithm \\t\" << std::endl;\n        std::cout << \"parameters age max \" << age_max  << std::endl;\n        std::cout << \"parameters clean_step \" << clean_step  << std::endl;\n        std::cout << \"parameters clean_ratio \" << clean_ratio  << std::endl;\n        std::cout << \"parameters num_threads \" << num_threads  << std::endl;\n    }\n    auto start = std::chrono::system_clock::now();\n    // Main algorithm\n    for (unsigned k = 0; k < num_threads; ++k) {\n        cntxt.__gen.seed(static_cast<uint64_t>(seeds[2*k]) << 32 | seeds[2*k+1]);\n        auto & counter_k = counters[k];\n        auto & pop_k = pops[k];\n        if (verbose) std::cout << pop_k.individuals.size() << std::endl;\n\n        threads[k] = std::thread(\n                [=, &pop_k, &counter_k]() mutable {\n                counter_k = popsim_algorithm(pop_k, cntxt, events,\n                        T0, T, clean_step, clean_ratio);\n                });\n    }\n    for (auto & th : threads) th.join();\n    auto end = std::chrono::system_clock::now();\n    auto elapsed_seconds = end-start;\n    double secs_main_algorithm = elapsed_seconds.count();\n\n    pop = merge(pops);\n    counter C {0, 0, 0 };\n    for (auto & Cs : counters) C += Cs;\n\n    // ***** OUTPUT_POPULATION\n    if (verbose) std::cout << \"Before output conversion \" << std::endl;\n\tstd::vector<double> _birth_date_out;\n\t_birth_date_out.reserve(pop.individuals.size()+pop.dead_individuals.size());\n\tstd::vector<double> _death_date_out;\n\t_death_date_out.reserve(pop.individuals.size()+pop.dead_individuals.size());\n\tif (verbose) std::cout << \"individuals size : \" << pop.individuals.size() << std::endl;\n\tfor (unsigned _k = 0; _k < pop.individuals.size(); ++_k) {\n\t    _birth_date_out.push_back( pop.individuals[_k].birth_date );\n\t    _death_date_out.push_back( pop.individuals[_k].death_date );\n\t    }\n\tif (verbose) std::cout << \"dead_individuals size : \" << pop.dead_individuals.size() << std::endl;\n\tfor (unsigned _k = 0; _k < pop.dead_individuals.size(); ++_k) {\n\t    _birth_date_out.push_back( pop.dead_individuals[_k].birth_date );\n\t    _death_date_out.push_back( pop.dead_individuals[_k].death_date );\n\t    }\n\tRcpp::DataFrame new_pop_df = Rcpp::DataFrame::create(\n\t    Rcpp::Named(\"birth\") = _birth_date_out,\n\t    Rcpp::Named(\"death\") = _death_date_out);\n\tif (verbose) std::cout << \"After output conversion \" << std::endl;\n\t\n    // *****\n\n    Rcpp::NumericVector logs = Rcpp::NumericVector::create(\n            Rcpp::Named(\"proposed_events\")= C.proposedEvents,\n            Rcpp::Named(\"effective_events\")= C.effectiveEvents,\n            Rcpp::Named(\"cleanall_counter\")= C.cleanAll,\n            Rcpp::Named(\"duration_main_algorithm\")=secs_main_algorithm);\n    // Returns population and simulation info\n    Rcpp::List L = Rcpp::List::create(\n            Rcpp::Named(\"population\") = new_pop_df,\n            Rcpp::Named(\"logs\") = logs);\n    return L;\n};\n"
#> 
#> $sourceCpp
#> $sourceCpp$functions
#> [1] "popsim_cpp"
#> 
#> $sourceCpp$modules
#> character(0)
#> 
#> $sourceCpp$cppSourcePath
#> [1] "/tmp/Rtmp0wvX9Z/sourceCpp-x86_64-pc-linux-gnu-1.0.9/file27266bfda486.cpp"
#> 
#> $sourceCpp$buildDirectory
#> [1] "/tmp/Rtmp0wvX9Z/sourceCpp-x86_64-pc-linux-gnu-1.0.9/sourcecpp_27267c9d8005"
#> 
#> 
#> $popsim_cpp
#> function (pop_df, events_bounds, parameters, age_max, T0, T, 
#>     clean_step, clean_ratio, seed, multithreading, num_threads = NA_integer_, 
#>     verbose = FALSE) 
#> .Call(<pointer: 0x7f05683ddbd0>, pop_df, events_bounds, parameters, 
#>     age_max, T0, T, clean_step, clean_ratio, seed, multithreading, 
#>     num_threads, verbose)
#> 
#> attr(,"class")
#> [1] "model"
# }